Добавили работу с комнатами
